final kredo class

+++ classes/user.php
@@ -1,7 +1,6 @@
 <?php
 require_once "database.php";
 
-
 class User extends Database
 {
     public function login($email, $password)
@@ -12,13 +11,13 @@ class User extends Database
         //1.If the email exists
         //2.If the password is correct.Compare the login password to database
 
-        echo "start serching<br>";
+        // echo "start serching<br>";
         if ($result->num_rows == 1) { //もし同じユーザー名があれば
             //email is exist 
             $user_details = $result->fetch_assoc();
-            echo $password . "<br>";
-            echo $user_details['password'] . "<br>";
-            echo $result->num_rows . "<br>";
+            // echo $password . "<br>";
+            // echo $user_details['password'] . "<br>";
+            // echo $result->num_rows . "<br>";
             // while ($user_details = $result->fetch_assoc()) { //user名はuniqueではないのでループで回す
 
             if (password_verify($password, $user_details['password'])) { //ユーザー名とパスワードが一致
@@ -47,80 +46,82 @@ class User extends Database
         }
     }
 
+    // change to snake case
     public function createUser($newFirstName, $newLastName, $newUserName, $newEmail, $newImage, $newImageTmp, $newPassword)
     {
-        //It need because fullname and email are unique.
-        $sql1 = "SELECT id,first_name,last_name,username,email,`image` from users WHERE email =  $newEmail "; //database information
-        $sql2 = "SELECT id,first_name,last_name,username,email,`image` from users WHERE first_name =  $newFirstName AND last_name = $newLastName "; //database information
+        //check unique the email 
+        $sql1 = "SELECT id from users WHERE email = '$newEmail';";; //database information
+        $result_sql1 = $this->conn->query($sql1);
+        if ($result_sql1->num_rows >= 1) {
+            header('Location: ../views/createUser.php?err=2');
+            exit;
+        }
 
-        // if ($this->conn->query($sql1)->num_rows >= 1) {//check unique about email,J:emailはユニーク設定だから結果が１以上あれば追加しない
-        //     die('The email address was used.');
-        // }
-        // if ($this->conn->query($sql2)->num_rows >= 1) {//check unique about fullname
-        //     die('The fullname was used.');
-        // }
 
-        if ($newImageTmp == null) { //user追加のSQL文
-            // $newImage = "../images/default.jpg";//69 and didn't work
-            // $_SESSION['icon']
+        // make sql sentence
+        if ($newImageTmp == null) { //if newImage is null
             $sql3 = "INSERT INTO users(first_name,last_name,username,email,`password`) VALUES ('$newFirstName','$newLastName','$newUserName','$newEmail','$newPassword')";
-        } else {
+        } else { //newImage is not null
             $sql3 = "INSERT INTO users(first_name,last_name,username,email,`image`,`password`) VALUES ('$newFirstName','$newLastName','$newUserName','$newEmail','$newImage','$newPassword')";
             $destination = "../images/" . basename($newImage);
             move_uploaded_file($newImageTmp, $destination); //if i didn't select a picture,the function didn't work
         }
+
+        //insert into database(users)
         if ($this->conn->query($sql3)) {
             //The sql for session
             $sql4 = "SELECT id,first_name,last_name,username,email,`image`,`password` FROM users WHERE username = '$newUserName'";
             $result = $this->conn->query($sql4);
             $user_details = $result->fetch_assoc();
 
+            //SESSION start
             session_start();
             $_SESSION['username'] = $newUserName;
             $_SESSION['user_id'] = $user_details['id'];
             $_SESSION['username'] = $user_details['username'];
-            // $_SESSION['fullname'] = $user_details['fullname'];
             $_SESSION['first_name'] = $user_details['first_name'];
             $_SESSION['last_name'] = $user_details['last_name'];
             $_SESSION['icon'] = $user_details['image'];
             $_SESSION['email'] = $user_details['email'];
             $_SESSION['password'] = $user_details['password'];
 
+            //this can show my post
             $id = $user_details['id'];
-
+            //follow myself
             $sql6 = "INSERT INTO follow(`user_id`,`following`) VALUES ($id,$id);";
             if ($this->conn->query($sql6)) {
                 header("location: ../views/home.php");
                 exit;
             } else {
-                die('Error inserting follow:' . $this->conn->error);
+                header('Location: ../views/createUser.php');
+                exit;
             }
         } else {
-            die("Error create user: " . $this->conn->error);
+            header('Location: ../views/createUser.php');
+            exit;
         }
     }
 
-
+    //editUser and snake case
     public function edit($editFirstName, $editLastName, $editUserName, $editEmail, $editImage, $editImageTmp)
     {
         session_start();
-        //!!!!!!!!!check!!!!!!!!!!!!!!!!
+
         //it need because SQL does't use $_SESSION[];and get id of the user
         $nowemail = $_SESSION['email'];
         //the user information now
         $sql = "SELECT id,first_name,last_name,username,email,`image` from users WHERE email =  '$nowemail' ";
-        //check same fullname;
-        $sql2 = "SELECT id,first_name,last_name,username,email,`image` from users WHERE first_name = '$editFirstName' AND last_name = '$editLastName';"; //check same fullname;
-        //check same email;
-        $sql3 = "SELECT id,first_name,last_name,username,email,`image` from users WHERE email = '$editEmail' ;";
-        //if those return bigger than 1, there is same fullname or email address.
-        $checkfullname = $this->conn->query($sql2);
+        //check unique email sentence
+        $sql3 = "SELECT id from users WHERE email = '$editEmail' ;";
+
+        //if those return bigger than 1, there is same email address.
         $checkemail = $this->conn->query($sql3);
-        if ($checkfullname->num_rows > 1) {
-            die('This fullname was used');
-        }
-        if ($checkemail->num_rows > 1) {
-            die('This eamil was used');
+        if ($checkemail->num_rows >= 1) {
+            if ($nowemail == $editEmail) {
+            } else {
+                header('Location: ../views/editUser.php?err=1');
+                exit;
+            }
         }
 
         //get id of the user and update database of the user
@@ -130,13 +131,15 @@ class User extends Database
 
             if ($editImageTmp == null) { //if image is null,not change one
                 $sql4 = "UPDATE users SET first_name = '$editFirstName',last_name ='$editLastName', username = '$editUserName', email = '$editEmail' WHERE id = $id;";
-            } else {
+            } else { //edit all of them
                 $sql4 = "UPDATE users SET first_name = '$editFirstName',last_name ='$editLastName', username = '$editUserName', email = '$editEmail',`image`='$editImage' WHERE id = $id;";
                 $destination = "../images/" . basename($editImage);
                 move_uploaded_file($editImageTmp, $destination); //if i didn't select a picture,the function didn't work
             }
 
+            //update database(users)
             if ($this->conn->query($sql4)) {
+                //The sentence for resave session
                 $sql5 = "SELECT id,first_name,last_name,username,email,`image` FROM users WHERE id = $id;";
                 $select = $this->conn->query($sql5);
                 $now_user_details = $select->fetch_assoc();
@@ -149,28 +152,27 @@ class User extends Database
                 header("Location: ../views/home.php");
                 exit;
             } else {
-                die("Error Updating users " . $this->conn->error);
+                header('Location: ../views/home.php');
+                exit;
             }
         } else {
-            die('Error select information of the user');
+            header('Location: ../views/home.php');
+            exit;
         }
     }
 
+    //search follow username
     public function searchUser($username)
     {
         $sql = "SELECT id,first_name,last_name,username,`image` from users WHERE username LIKE '%$username%';";
         if ($result = $this->conn->query($sql)) {
-
-            // foreach ($result as $number) {
-            //     $user[] =$number['username'] . "<br>";
-            // }
-            // return $user;
             return $result;
         } else {
             die("ERROR:searching username");
         }
     }
 
+    //getUserDetail
     public function getUser($id)
     {
         $sql = "SELECT * from users WHERE id = $id";
@@ -182,28 +184,26 @@ class User extends Database
         }
     }
 
-
-    public function getUserImage($id)
-    {
-        $sql = "SELECT `image` from users WHERE id = $id;";
-
-        if ($result = $this->conn->query($sql)) {
-            return $result->fetch_assoc();
-        } else {
-            die('Error selecting userImage  : ' . $this->conn->error);
-        }
-    }
-
+    //delete User
     public function deleteUser($id)
     {
         $sql = "DELETE FROM users WHERE id = $id";
         if ($this->conn->query($sql)) {
-            $sql2 = "DELETE FROM follow WHERE `following` = $id";
+            $sql2 = "DELETE FROM follow WHERE `following` = $id OR `user_id` = $id;";
             if ($this->conn->query($sql2)) {
-                session_unset();
-                session_destroy();
-                header('Location: ../views/index.php');
-                exit;
+                $sql3 = "DELETE FROM posts WHERE `user_id` = $id";
+                if ($this->conn->query($sql3)) {
+                    $sql4 = "DELETE FROM chat WHERE `user_id` = $id";
+                    if ($this->conn->query($sql4)) {
+
+                        session_unset();
+                        session_destroy();
+                        header('Location: ../views/index.php');
+                        exit;
+                    }
+                }
+            } else {
+                die('Error deleting user' . $this->conn->error);
             }
         } else {
             die('Error deleting user' . $this->conn->error);
